moved get_next_care_date function to utils folder
